brett_projekt v0.0.1 BrettProjekt.Game

Functions

add_new_player(game, name)

Add a new player with a given name to the game.

The player name will be trimmed and converted to lowercase.

Arguments

game: The game to which the player is going to be added

name: The name of the new player

create(game_id)

Creates a new game with id.

Parameters

game_id: a unique identifier of the game. Normally generated by the game manager.

Examples

BrettProjekt.Game.create System.unique_integer

disable_join(game)

Disables joining of new players

Arguments

game: The game on which joining of new players should be disabled

get_id(game)

Returns the unique game-id set when calling create/1

get_players(game)

Returns a map of players with the player-id as key and a player-struct as the value.

has_player?(game, player_name)

Checks if a player with name is in the game

The name that will be checked gets trimmed and converted to lowercase.

Arguments

game: The game to check for players

player_name: Name of the player to search for

join_enabled?(game)

Returns whether a game can be joined by a new player

Parameters

game: The game to check on

Examples

iex> {:ok, game} = BrettProjekt.Game.create System.unique_integer # Should be generated by the GameManager
iex> BrettProjekt.Game.join_enabled? game
true
iex> BrettProjekt.Game.disable_join game
iex> BrettProjekt.Game.join_enabled? game
false

startable?(game)