brett_projekt v0.0.1 BrettProjekt.GameManager

Provides state and functions for handling Game instances.

The GameManager is implemented as a GenServer holding different Game instances. These can be accessed by convenient functions.

It also provides functions for generating unique game-identifiers.

Functions

add_game(game_manager, game)

Adds an existing BrettProjekt.Game to the GameManager

The BrettProjekt.Game needs to have a unique id already set. Get it form generate_game_id/1.

Examples

iex> {:ok, game_manager} = BrettProjekt.GameManager.start_link
iex> {:ok, game} = BrettProjekt.Game.create(BrettProjekt.GameManager.generate_game_id game_manager)
iex> {:ok, game_id, ^game} = BrettProjekt.GameManager.add_game game_manager, game
iex> is_binary game_id
true

add_new_game(game_manager)

Creates a BrettProjekt.Game and adds it to the GameManager

A unique game-id will automatically be generated.

Examples

iex> {:ok, game_manager} = BrettProjekt.GameManager.start_link
iex> {:ok, _game_id, game} = BrettProjekt.GameManager.add_new_game game_manager
iex> is_pid game
true

generate_game_id(game_manager)

Generates a unique game-id string

The resulting string will be a Base-32 encoded 6-character long string.

get_game_by_id(game_manager, game_id)

Finds a game by id

Returns nil no game with given id is registered.

Arguments

game_manager: The game_manager to search on

game_id: The game_id to search for

Example

iex> {:ok, game_manager} = BrettProjekt.GameManager.start_link
iex> {:ok, game_id, _} = BrettProjekt.GameManager.add_new_game game_manager
iex> game1 = BrettProjekt.GameManager.get_game_by_id game_manager, game_id
iex> is_pid game1
true
iex> game2 = BrettProjekt.GameManager.get_game_by_id game_manager, "foobar"
iex> is_nil game2
true

get_games(game_manager)

Returns a map of registered games

The resulting map has game-ids as keys and game-pids as values.

start_link()

Starts a new GameManager instance

The GenServer will be started with no name assigned.

Example

{:ok, game_manager} = BrettProject.GameManager.start_link

start_link(name)

Starts a new GameManager instance with a name assigned

Arguments

name: Name of the GameManager (as an Atom)

Examples

{:ok, game_manager} = BrettProjekt.GameManager.start_link :my_fancy_game_manager