brett_projekt v0.0.1 BrettProjekt.Question.ServerManager

Manager for handling BrettProjekt.Question.Server easily.

The BrettProjekt-Game requires a question-server to be running and loaded with questions to ask the players. This module creates a server and loads the questions into it. (load_questions_from_file/2)

A listener for file-changes can be enabled, so that a new question-server is created with the changed set of questions.

When starting a game, the current BrettProjekt.Question.Server process should be retrieved using get_question_server/1 and stored in the game state. Therefore, if a new set of questions gets loaded, existing games will keep using the original question-set but new games start with an up-to-date question-set.

Functions

get_question_server(server_manager)

Retrieves the current question-server

This value needs to be cached per Game to prevent changing the question-set (including ids, etc.)

Returns

{:error, :no_question_server}: No question-server has been started yet.

{:ok, pid_of_question_server}

load_questions_from_file(server_manager, file)

Loads questions from a JSON-file, parses it and (on success) replaces the current question-server

Arguments

file: The file to load (absolute or relative path)

Returns

:ok

{:error, :file_not_found}

{:error, :file_read_error}

{:error, :json_decoding_error}

{:error, :file_version_not_supported}: No parser for the file-format version available

{:error, :file_invalid}: Error while parsing the questions file

start_link()

Start the ServerManager without a name

Examples

iex> {:ok, server_manager} = BrettProjekt.Question.ServerManager.start_link
iex> is_pid server_manager
true

start_link(name)

Start the ServerManager under a specific name (atom)

Examples

iex> {:ok, server_manager} = BrettProjekt.Question.ServerManager.start_link :my_server_manager
iex> is_pid server_manager
true

start_link(name, filename)

Starts the ServerManager with a name and loads an initial questions-file

To start it without a name assinged, pass nil