brett_projekt v0.0.1 BrettProjekt.Web.GameChannel

Functions

auth_token_valid?(auth_token, game_id)

Check validity of auth_token.

Check whether

game_id is the same as the one in the auth_token

auth_token is not nil

auth_token has not expired

auth_token is not invalid

Examples

iex> game_id = "abc"
iex> token = Phoenix.Token.sign(
         BrettProjekt.Web.Endpoint,
         "user_auth",
         %{game_id: game_id, player_id: 1}
     )
iex> BrettProjekt.Web.GameChannel.auth_token_valid? token, game_id

broadcast_lobby_update(socket, game)

Send a broadcast to all users in this game/channel with the updated list of players and startable status of the game.

code_change(old, socket, extra)

Callback implementation for Phoenix.Channel.code_change/3.

handle_in(arg1, payload, socket)

Sets the team membership of the user with the “auth_token” in the payload. Broadcasts a lobby_update to all users in this game/channel.

Returns

{:reply, :ok, socket}

{:reply, {:error, %{reason: msg}}, socket}

handle_info(message, state)

Callback implementation for Phoenix.Channel.handle_info/2.

join(channel, payload, socket)

Join the channel of a particular game.

payload has to contain an “auth_token” valid for the socket

Returns

{:ok, socket}

{:error, msg} msg is an error code from auth_token_valid

terminate(reason, socket)

Callback implementation for Phoenix.Channel.terminate/2.