Hasbihal v0.0.1 HasbihalWeb.Router.Helpers

Module with named helpers generated from HasbihalWeb.Router.

Functions

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

room_path(conn_or_endpoint, action, id)

room_path(conn_or_endpoint, action, id, params)

room_url(conn_or_endpoint, action, id)

room_url(conn_or_endpoint, action, id, params)

session_path(conn_or_endpoint, action)

session_path(conn_or_endpoint, action, params)

session_path(conn_or_endpoint, action, id, params)

session_url(conn_or_endpoint, action)

session_url(conn_or_endpoint, action, params)

session_url(conn_or_endpoint, action, id, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, id)

user_path(conn_or_endpoint, action, id, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, id)

user_url(conn_or_endpoint, action, id, params)