brett_projekt v0.0.1 BrettProjekt.GameManager

Provides state and functions for handling Game instances.

The GameManager is implemented as an Agent holding different Game instances. These can be accessed by convenient functions.

It also provides functions for generating unique game-identifiers.

Types

game()

game() :: pid

game_id()

game_id() :: String.t

Functions

add_new_game(game_manager)

add_new_game(Agent.agent) :: {:ok, game, game_id}

Creates a BrettProjekt.Game and adds it to the GameManager

A unique game-id will automatically be generated.

Examples

iex> {:ok, game_manager} = BrettProjekt.GameManager.start_link
iex> {:ok, _game_id, game} = BrettProjekt.GameManager.add_new_game game_manager
iex> is_pid game
true

get_game_by_id(game_manager, game_id)

get_game_by_id(Agent.agent, game_id) :: game | nil

start_link()

Starts a new GameManager instance

The Agent will be started with no name assigned.

Example

{:ok, game_manager} = BrettProject.GameManager.start_link

start_link(name)

Starts a new GameManager instance with a name assigned

Arguments

name: Name of the GameManager (as an Atom)

Examples

{:ok, game_manager} = BrettProjekt.GameManager.start_link :my_fancy_game_manager